Revert "node_deletion flag added, extra logs added"
This reverts commit b4f08f31009ae115254bf6efbfc4a6369d595efa.
