+++ keep/api/utils/import_ee.py
@@ -17,8 +17,7 @@ if EE_ENABLED:
     sys.path.insert(0, path_with_ee)
 
     from ee.experimental.incident_utils import mine_incidents_and_create_objects, generate_update_incident_summary, generate_update_incident_name  # noqa
-    from ee.experimental.generative_utils import generate_incident_summary, generate_incident_name, SUMMARY_GENERATOR_VERBOSE_NAME, NAME_GENERATOR_VERBOSE_NAME  # noqa
-    from ee.experimental.incident_utils import ALGORITHM_VERBOSE_NAME  # noqa 
+    from ee.experimental.incident_utils import ALGORITHM_VERBOSE_NAME, SUMMARY_GENERATOR_VERBOSE_NAME, NAME_GENERATOR_VERBOSE_NAME  # noqa 
 else:
     mine_incidents_and_create_objects = NotImplemented
     generate_update_incident_summary = NotImplemented
